Refactored PageMetadata HOC
